Corrected comments regarding HomeworkDay1.

+++ src/HomeworkDay1/OddNumber.java
@@ -9,7 +9,7 @@ public class OddNumber {
     public static void main(String[] args){
         Scanner scanner = new Scanner(System.in);
         System.out.println("Enter number:");
-        double n = scanner.nextDouble();
+        int n = scanner.nextInt();
 
         if (n % 2==0) {
             System.out.println("Even Number");
